test: add upgrade test for enterprise DA
