+++ tests/pr_test.go
@@ -50,18 +50,6 @@ func setupOptions(t *testing.T, prefix string, dir string) *testhelper.TestOptio
 	return options
 }
 
-func TestRunUpgradeExample(t *testing.T) {
-	t.Parallel()
-
-	options := setupOptions(t, "event-streams-upg", completeExampleTerraformDir)
-
-	output, err := options.RunTestUpgrade()
-	if !options.UpgradeTestSkipped {
-		assert.Nil(t, err, "This should not have errored")
-		assert.NotNil(t, output, "Expected some output")
-	}
-}
-
 func TestRunQuickstartSolution(t *testing.T) {
 	t.Parallel()
 
@@ -85,12 +73,10 @@ func TestRunQuickstartSolution(t *testing.T) {
 	assert.NotNil(t, output, "Expected some output")
 }
 
-func TestEnterpriseSolutionInSchematics(t *testing.T) {
-	t.Parallel()
-
+func setupEnterpriseOptions(t *testing.T, prefix string) *testschematic.TestSchematicOptions {
 	options := testschematic.TestSchematicOptionsDefault(&testschematic.TestSchematicOptions{
 		Testing:            t,
-		Prefix:             "es-ent",
+		Prefix:             prefix,
 		BestRegionYAMLPath: regionSelectionPath,
 		TarIncludePatterns: []string{
 			"*.tf",
@@ -120,43 +106,25 @@ func TestEnterpriseSolutionInSchematics(t *testing.T) {
 		{Name: "access_tags", Value: permanentResources["accessTags"], DataType: "list(string)"},
 		{Name: "resource_tags", Value: options.Tags, DataType: "list(string)"},
 	}
+	return options
+}
 
+// Test for the Enterprise DA
+func TestEnterpriseSolutionInSchematics(t *testing.T) {
+	t.Parallel()
+
+	options := setupEnterpriseOptions(t, "es-ent")
 	err := options.RunSchematicTest()
 	assert.Nil(t, err, "This should not have errored")
 }
 
-func TestFSCloudInSchematics(t *testing.T) {
+// Upgrade test for the Enterprise DA
+func TestRunUpgradeEnterpriseDA(t *testing.T) {
 	t.Parallel()
 
-	options := testschematic.TestSchematicOptionsDefault(&testschematic.TestSchematicOptions{
-		Testing:            t,
-		Prefix:             "es-fscloud",
-		BestRegionYAMLPath: regionSelectionPath,
-		TarIncludePatterns: []string{
-			"*.tf",
-			fsCloudTerraformDir + "/*.tf",
-			"modules/fscloud/*.tf",
-		},
-		/*
-			Comment out the 'ResourceGroup' input to force this tests to create a unique resource group to ensure tests do
-			not clash. This is due to the fact that an auth policy may already exist in this resource group since we are
-			re-using a permanent HPCS instance and a permanent Event Streams instance. By using a new resource group, the auth policy will not already exist
-			since this module scopes auth policies by resource group.
-		*/
-		//ResourceGroup:      resourceGroup,
-		TemplateFolder:         fsCloudTerraformDir,
-		Tags:                   []string{"test-schematic"},
-		DeleteWorkspaceOnFail:  false,
-		WaitJobCompleteMinutes: 180,
-	})
-
-	options.TerraformVars = []testschematic.TestSchematicTerraformVar{
-		{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
-		{Name: "prefix", Value: options.Prefix, DataType: "string"},
-		{Name: "kms_key_crn", Value: permanentResources["hpcs_south_root_key_crn"].(string), DataType: "string"},
-		{Name: "event_streams_source_crn", Value: permanentResources["event_streams_us_south_crn"].(string), DataType: "string"},
+	options := setupEnterpriseOptions(t, "ev-st-upg")
+	err := options.RunSchematicUpgradeTest()
+	if !options.UpgradeTestSkipped {
+		assert.Nil(t, err, "This should not have errored")
 	}
-
-	err := options.RunSchematicTest()
-	assert.Nil(t, err, "This should not have errored")
 }
